# cmd/api/docs/swagger.yaml
definitions:
...
  helperStruct.Coupons:
    properties:
      code:
        type: string
      discountmaximumamount:
        type: number
      discountpercent:
        type: number
      expirationdate:
        type: string
      minimumpurchaseamount:
        type: number
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /addcoupontocart/{coupon_id}:
    patch:
      consumes:
      - application/json
      description: User can add coupon to the cart
      operationId: applay-coupon-to-cart
      parameters:
      - description: coupon_id
        in: path
        name: coupon_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can add a coupon to the cart
      tags:
      - Cart
  /addtocart/{product_item_id}:
    post:
      consumes:
      - application/json
      description: User can add product item to the cart
      operationId: add-to-cart
      parameters:
      - description: product_item_id
        in: path
        name: product_item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can add a product item to the cart
      tags:
      - Cart
...
  /admin/createcoupon/:
    post:
      consumes:
      - application/json
      description: Admin can create new coupons
      operationId: create-coupon
      parameters:
      - description: details of new coupon to be created
        in: body
        name: new_coupon_details
        required: true
        schema:
          $ref: '#/definitions/helperStruct.Coupons'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can create new coupon
      tags:
      - Coupon
  /admin/deletecoupon/{couponId}:
    delete:
      consumes:
      - application/json
      description: Admin can delete existing coupon
      operationId: delete-coupon
      parameters:
      - description: details of coupon to be updated
        in: path
        name: coupon_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can delete existing coupon
      tags:
      - Coupon
...
  /admin/updatecoupen/{couponId}:
    patch:
      consumes:
      - application/json
      description: Admin can update existing coupon
      operationId: update-coupon
      parameters:
      - description: Coupon ID
        in: path
        name: couponId
        required: true
        type: string
      - description: details of coupon to be updated
        in: body
        name: coupon_details
        required: true
        schema:
          $ref: '#/definitions/helperStruct.Coupons'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can update existing coupon
      tags:
      - Coupon
  /admin/viewcoupon/{couponId}:
    get:
      consumes:
      - application/json
      description: Admins and users can see coupon with id
      operationId: view-coupon-by-id
      parameters:
      - description: coupon_id
        in: path
        name: coupon_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admins and users can see coupon with coupon id
      tags:
      - Coupon
  /admin/viewcoupons/:
    get:
      consumes:
      - application/json
      description: Admins and users can see all available coupons
      operationId: view-coupons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admins and users can see all available coupons
      tags:
      - Coupon
...
  /listcart:
    get:
      consumes:
      - application/json
      description: User can view cart and cart items
      operationId: view-cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can view cart items and total
      tags:
      - Cart
...
  /removecoupon:
    patch:
      consumes:
      - application/json
      description: User can add coupon to the cart
      operationId: remove-coupon-to-cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can remove the coupon that added to the cart
      tags:
      - Cart
  /removefromcart/{product_item_id}:
    delete:
      consumes:
      - application/json
      description: User can remove product from cart
      operationId: remove-from-cart
      parameters:
      - description: product_item_id
        in: path
        name: product_item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Remove a product from the cart
      tags:
      - Cart
...
swagger: "2.0"